requirements.txt
asgiref==3.3.2
Django==3.2
djangorestframework==3.12.4
django-filter==2.4.0
pytz==2021.1
sqlparse==0.4.1
